add dpsTransExecutor
git-svn-id: http://192.168.20.11/sequoiadb/trunk@37756 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/dps/dpsTransExecutor.cpp
@@ -178,7 +178,7 @@ namespace engine
       _transTimeout = timeout ;
    }
 
-   UINT32 _dpsTransExecutor::getTransTimeout() const
+   INT64 _dpsTransExecutor::getTransTimeout() const
    {
       return _transTimeout ;
    }
